web_gateway: carve agent_card.rs (pure move)
Advertisement + the agent card move to web_gateway/agent_card.rs:
advertise-URL resolution/auto-detection, the local agent card, and
gateway config assembly, with their 16 tests.
